合并dev

+++ src/main/java/org/nf/neoflow/models/ModelNode.java
@@ -55,10 +55,10 @@ public class ModelNode {
     private Integer location;
 
     @Schema(name = "x坐标")
-    private Float x;
+    private Double x;
 
     @Schema(name = "y坐标")
-    private Float y;
+    private Double y;
 
     /**
      * 复制忽略属性
